Moved session secret handling to new file
