+++ kit/score/registry_test.go
@@ -2,22 +2,11 @@ package score
 
 import (
 	"fmt"
-	"os"
 	"strings"
 	"testing"
 
 	"github.com/autograde/quickfeed/kit/score/testdata/a"
 )
-
-// To run this test, use this command:
-//   QUICKFEED_SESSION_SECRET=hei go test -v -run TestSessionSecret
-//
-func TestSessionSecret(t *testing.T) {
-	sessionSecret := os.Getenv(secretEnvName)
-	if sessionSecret != "" {
-		t.Fatalf("Unexpected access to %s=%s", secretEnvName, sessionSecret)
-	}
-}
 
 func fib() {}
